finance-app/backend/requirements.txt
fastapi
uvicorn[standard]
prisma
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
openpyxl
